requirements.txt
alembic==1.8.1
altair==4.2.0
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
blinker==1.5
cachelib==0.9.0
cachetools==4.2.1
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==38.0.4
decorator==5.1.1
distlib==0.3.6
dnspython==2.2.1
email-validator==1.1.3
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Email==1.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Minify==0.37
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==1.27.0
google-auth-oauthlib==0.4.2
greenlet==1.1.3.post0
gunicorn==20.1.0
htmlmin==0.1.12
httpagentparser==1.9.5
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
jsonschema==4.17.3
lesscpy==0.15.1
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.0.1
MechanicalSoup==1.1.0
mysql-connector-python==8.0.28
mysqlclient==2.1.1
numpy==1.23.4
oauthlib==3.1.0
openai==0.24.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.1
pandas-stubs==1.5.2.221124
phpserialize==1.3
Pillow==9.3.0
platformdirs==2.5.4
ply==3.11
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
rcssmin==1.1.1
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rich==12.6.0
rsa==4.7.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
streamlit==1.15.2
stripe==5.0.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
twilio==7.16.0
types-pytz==2022.5.0.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.8
validators==0.20.0
virtualenv==20.17.0
watchdog==2.2.0
Werkzeug==2.0.2
WTForms==3.0.0
xxhash==3.1.0
zipp==3.10.0
